Create xargs_sweep
As of this writing, this is my fastest class C ping sweep. I have yet to clock this at faster than one second. Then, again, this iteration may not adequately capture fractions of a second in the final print statement. I will have to look into this. The speeds I have clocked for now are remarkable.