Create xargsoneline
This is about as compacts as you can get with a sweep while still taking advantage of features such as parallelization. The downside to this is that it has no logging, no readout for dead hosts. I do take care of these things in the robust versions of sweeps included in this repository. 

This one liner is a great asset for a penetration tester who need a way to rapidly fire up a ping sweep without the need of a script handy. If the job has to be stealth, swift and effective, this code achieves that. 

I have even simpler syntax I will include but this is my favorite oneliner. I love that it runs the sweep in exact parallel execution given the specificity of xargs -P254, one process per host address means flawless timing and reliable output.