Update xargsoneline
I added a sort parameter as a final element to this pipe and I'm glad I did ... I have seen more than a few one liner sweeps that fail to sort the IP addresses. If you have only a small number of hosts to enumerate, it's not that important. 

However, as an IT professional or analyst you will surely come across hundreds of hosts - and you won't have the time or attention span at the end of a week to sift through the data. Anyway - this solves that. It's also very cool that sort has that "-V" option as I've seen many ways to do this, some of which are dang nasty.